requirements.txt
pandas
numpy
streamlit
flask
openai
python-dotenv
st_pages
streamlit-pages
